perf(components) add composables and constants, update components
