+++ src/constants/router-names.ts
@@ -1,0 +1,4 @@
+export default {
+    HOME: "home",
+    TODOLIST: "todoList"
+}
